Creating reusable funcs to manage lambdas
