+++ aws/lambda-container/deployer.go
@@ -4,22 +4,13 @@ import (
 	"context"
 	"fmt"
 	"github.com/nullstone-io/deployment-sdk/app"
-	nsaws "github.com/nullstone-io/deployment-sdk/aws"
 	"github.com/nullstone-io/deployment-sdk/aws/lambda"
-	"github.com/nullstone-io/deployment-sdk/docker"
+	nslambda "github.com/nullstone-io/deployment-sdk/aws/lambda"
 	env_vars "github.com/nullstone-io/deployment-sdk/env-vars"
 	"github.com/nullstone-io/deployment-sdk/logging"
 	"github.com/nullstone-io/deployment-sdk/outputs"
 	"gopkg.in/nullstone-io/go-api-client.v0"
 )
-
-type Outputs struct {
-	Region       string          `ns:"region"`
-	Deployer     nsaws.User      `ns:"deployer"`
-	LambdaArn    string          `ns:"lambda_arn"`
-	LambdaName   string          `ns:"lambda_name"`
-	ImageRepoUrl docker.ImageUrl `ns:"image_repo_url,optional"`
-}
 
 func NewDeployer(osWriters logging.OsWriters, nsConfig api.Config, appDetails app.Details) (app.Deployer, error) {
 	outs, err := outputs.Retrieve[Outputs](nsConfig, appDetails.Workspace)
@@ -48,15 +39,16 @@ func (d Deployer) Deploy(ctx context.Context, meta app.DeployMetadata) (string, 
 	}
 
 	fmt.Fprintf(stdout, "Updating lambda environment variables\n")
-	config, err := GetFunctionConfig(ctx, d.Infra)
+	config, err := nslambda.GetFunctionConfig(ctx, d.Infra)
 	if err != nil {
 		return "", fmt.Errorf("error retrieving lambda configuration: %w", err)
 	}
 	updates := lambda.MapFunctionConfig(config)
 	updates.Environment.Variables = env_vars.UpdateStandard(updates.Environment.Variables, meta)
-	if err := UpdateFunctionConfig(ctx, d.Infra, updates); err != nil {
+	if err := nslambda.UpdateFunctionConfig(ctx, d.Infra, updates); err != nil {
 		return "", fmt.Errorf("error updating lambda configuration: %w", err)
 	}
+
 	fmt.Fprintf(stdout, "Updated lambda environment variables\n")
 
 	fmt.Fprintf(stdout, "Updating lambda to %q\n", meta.Version)
