+++ aws/lambda-zip/outputs.go
@@ -1,6 +1,7 @@
 package lambda_zip
 
 import (
+	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/nullstone-io/deployment-sdk/aws"
 	"strings"
 )
@@ -17,10 +18,18 @@ type Outputs struct {
 	ArtifactsKeyTemplate string     `ns:"artifacts_key_template"`
 }
 
+func (o Outputs) FunctionName() string {
+	return o.LambdaName
+}
+
 func (o Outputs) ArtifactsKey(appVersion string) string {
 	tmpl := o.ArtifactsKeyTemplate
 	if tmpl == "" {
 		tmpl = "{{app-version}}"
 	}
 	return strings.Replace(tmpl, KeyTemplateAppVersion, appVersion, -1)
+}
+
+func (o Outputs) DeployerAwsConfig() aws.Config {
+	return nsaws.NewConfig(o.Deployer, o.Region)
 }
