+++ aws/lambda-zip/update_lambda_version.go
@@ -4,12 +4,10 @@ import (
 	"context"
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/lambda"
-	nsaws "github.com/nullstone-io/deployment-sdk/aws"
-	nslambda "github.com/nullstone-io/deployment-sdk/aws/lambda"
 )
 
 func UpdateLambdaVersion(ctx context.Context, infra Outputs, version string) error {
-	λClient := lambda.NewFromConfig(nsaws.NewConfig(infra.Deployer, infra.Region), nslambda.StandardRetrierFn)
+	λClient := lambda.NewFromConfig(infra.DeployerAwsConfig())
 	_, err := λClient.UpdateFunctionCode(ctx, &lambda.UpdateFunctionCodeInput{
 		FunctionName: aws.String(infra.LambdaName),
 		DryRun:       false,
